tools/requirements.txt
PyMuPDF>=1.24.0
anthropic>=0.40.0
python-dotenv>=1.0.0
Pillow>=10.0.0
